added asyncHandler to song/artist controllers in case of errors when querying spotify. Split controllers for song/artist into separate files.
