+++ src/queries/queries.controller.ts
@@ -3,11 +3,8 @@ import service from "./queries.service";
 import asyncHandler from "../errors/asyncHandler";
 import { Query } from "../utils/types";
 
-async function list(req: Request, res: Response, next: NextFunction) {
-    const { limit = 0 } = req.query;
-        if(isNaN(Number(limit))) { 
-            next({ status: 400, message: "Limit must be a number" })
-        }
+async function list(req: Request, res: Response) {
+    const { limit } = res.locals;
         const queries = await service.list(Number(limit));
         return res.json({ data: queries });
 }
@@ -18,7 +15,17 @@ async function post (req: Request, res: Response) {
     res.json({ data: await service.post({ search_keyword, artist_name, num_songs }) });
   } 
 
+
+function limitIsNumber(req: Request, res: Response, next: NextFunction) {
+    const { limit = 0 } = req.query;
+    if (isNaN(Number(limit))) {
+        next({ status: 400, message: "Limit must be a number" });
+    }
+    res.locals.limit = limit;
+    next();
+}  
+
  export default {
-    list: asyncHandler(list),
+    list: [limitIsNumber, asyncHandler(list)],
     post: asyncHandler(post),
  }
